Add files via upload
 Implement FAÇADE pattern for the following application: 

A member of a system is ordering some food from a store but system needs to verify username and password using a subclass. User also have privilege to use credit card for payment to purchase items. An application should be able to perform verification for credit card and payment amount (remaining/limit of a credit card).